refactor code, some progress on multiplayer support, although idrk how the server works much so not fully implemented, recoil, reloading, also muzzle flash but i dont think it works yet, ammo as well
