+++ src/hooks/useGameSocket.ts
@@ -1,0 +1,31 @@
+import { useState, useEffect, useRef } from 'react';
+
+export default function useGameSocket(url: string) {
+  const [connected, setConnected] = useState(false);
+  const wsRef = useRef<WebSocket | null>(null);
+
+  useEffect(() => {
+    const ws = new WebSocket(url);
+    wsRef.current = ws;
+
+    ws.onopen = () => {
+      console.log("Connected to game server");
+      setConnected(true);
+    };
+
+    ws.onclose = () => {
+      console.log("Disconnected from game server");
+      setConnected(false);
+    };
+
+    ws.onerror = (error) => {
+      console.error("WebSocket error:", error);
+    };
+
+    return () => {
+      ws.close();
+    };
+  }, [url]);
+
+  return { ws: wsRef.current, connected };
+}
